refactor: 개발 확장성 및 직관성을 위해 패키지 구조 정리
