+++ src/main/java/gift/dto/user/UserRequestDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
