+++ src/main/java/gift/model/Product.java
@@ -1,6 +1,6 @@
 package gift.model;
 
-import gift.validation.RestrictedKeyword;
+import gift.validation.annotation.RestrictedKeyword;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
