+++ src/main/java/gift/validation/annotation/RestrictedKeyword.java
@@ -1,5 +1,6 @@
-package gift.validation;
+package gift.validation.annotation;
 
+import gift.validation.validator.RestrictedKeywordValidator;
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
